requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=9.5.0
scikit-learn>=1.2.0

# MedMNIST dataset
medmnist>=2.1.0

# Transformers for VLM
transformers>=4.36.0
accelerate>=0.27.0

# For notebook
ipykernel>=6.25.0
ipywidgets>=8.1.0
tqdm>=4.65.0

# Hugging Face
huggingface_hub>=0.20.0